powl, ceill, floorl, fmodl -> std:: equivalents from <cmath>
value::to_integral_number works with std::pow etc, fixes
build on FreeBSD, which doesn't have these in the system C headers,
but the system C++ compiler (GCC) has them.

* libflusspferd/spidermonkey/value.cpp

+++ libflusspferd/spidermonkey/value.cpp
@@ -103,9 +103,9 @@ double value::to_integral_number(int bits, bool signedness) const {
   if (!std::isfinite(value))
     return 0;
 #endif
-  long double maxU = powl(2, bits);
-  value = value < 0 ? ceill(value) : floorl(value);
-  value = fmodl(value, maxU);
+  long double maxU = std::pow(2.0L, bits);
+  value = value < 0 ? std::ceil(value) : std::floor(value);
+  value = std::fmod(value, maxU);
   if (value < 0)
     value += maxU;
   if (signedness) {
